+++ changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `install-boxlang` refactoring to be similar to all scripts.
+
 ## [1.11.0] - 2025-06-11
 
 ### Added